+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * SIGKILL for second attempt to kill Factorio when ending a debug session
 * Fix rawlen() fix incorrectly applied to LuaCustomTable
 * Add adjustMods and disableExtraMods to launch args
+* Step inside zip files with slevesque.vscode-zipexplorer
 
 ## 0.18.14